This fixes an issue with StrictRedis and zadd.
The redis-py StrictRedis connection class adheres to the actual
redis API definition. Usually this is not an issue as redis-py
Redis class methods have the same list of arguments.

The issue arises when using ZADD, since in redis it expects
the ordered set name and the items to add with score and key in this
order.

StrictRedis uses the correct order of arguments, whereas Redis inverts
key and score. This gives users an exception if they use Retools Limiter
as it is implemented today with a StrictRedis connection.