Removed setup_requires nose
I removed it because it becomes a dependency whenever installing retools in production. This would not be an issue if there was not a bug in setuptools.

If there's any packages in setup_requires, setuptools downloads it without respecting any proxy configuration. Since many production servers block outbound connections, it's desirable to remove this attribute.